+++ src/lib/molecules/DatePickerInput/DateRangeInput.js
@@ -10,7 +10,7 @@ import { useClick, useDeviceType, useKeyDown } from 'hooks';
 import { dayjsWithPlugins } from 'wrappers';
 
 // Components
-import Popover from '../../atoms/PopoverV2';
+import Popover from '../../atoms/Popover';
 import ExtendedInput from '../ExtendedInput';
 import DatePicker from '../../organisms/DatePicker';
